+++ NEWS.md
@@ -6,7 +6,7 @@ This version of `tern` introduces a major rewriting of `tern` due to the change 
 * Summarize variables: S3 generic `s_summary` (with numeric, factor and character methods), `summarize_vars`.
 * Estimate difference in terms of responder proportions: `estimate_proportion_diff`, `s_proportion_diff`.
 * Test the difference in responder proportions: `prop_chisq`, `prop_cmh`.
-* Estimate difference in terms of Odds Ratio: `or_glm`.
+* Estimate difference in terms of Odds Ratio: `s_odds_ratio`, `or_glm`.
 
 * issues with no news: